+++ src/plugins/widgets/bookmarks/types.ts
@@ -1,15 +1,19 @@
 import { API } from "../../types";
 
 type Data = {
-  rootBookmark: string;
+  rootBookmark: string | null;
   maxWidth: number;
+  maxHeight: number;
   wrap: boolean;
+  navigationStyle: 'drill-down' | 'expand-collapse';
 };
 
 export type Props = API<Data>;
 
 export const defaultData: Data = {
-  rootBookmark: "",
+  rootBookmark: null,
   maxWidth: 30,
+  maxHeight: 40,
   wrap: true,
+  navigationStyle: 'drill-down'
 };